pyproject.toml
[project]
name = "car-api"
version = "0.1.0"
description = ""
authors = [
    {name = "GenshinEnjoyer32",email = "iddmitriev@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
